+++ ompi/mca/ptl/sm/ptl_sm_component.c
@@ -247,8 +247,8 @@ mca_ptl_base_module_t** mca_ptl_sm_component_init(
 #if OMPI_ENABLE_PROGRESS_THREADS == 1
     /* create a named pipe to receive events  */
     sprintf(mca_ptl_sm_component.sm_fifo_path, 
-        "%s/sm_fifo.%d", orte_process_info.job_session_dir,
-         orte_process_info.my_name->vpid);
+        "%s/sm_fifo.%lu", orte_process_info.job_session_dir,
+         (unsigned long)orte_process_info.my_name->vpid);
     if(mkfifo(mca_ptl_sm_component.sm_fifo_path, 0660) < 0) {
         opal_output(0, "mca_ptl_sm_component_init: mkfifo failed with errno=%d\n",errno);
         return NULL;
